api/requirements.txt
django>=5.1
djangorestframework>=3.15
psycopg>=3.1
psycopg-binary>=3.2
python-environ>=0.4
django-vectortiles>=1.0